Add files via upload
This code came from refactoring Brad Traversy's (much thanks to him!) Expense Tracker project for React Hooks with Context API YouTube video.  It is not yet fully complete, but it renders successfully and will be the Score Sheet component for my new version of the game of chess that I invented about 11 years ago originally, with only a hard copy mail-order version.